// src/hooks/orval/model/tournamentsRecord.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * PocketBase API
 * API documentation generated from PocketBase schema with collection grouping
 * OpenAPI spec version: 1.0.0
 */
import type { TournamentsRecordBracketFormat } from './tournamentsRecordBracketFormat.ts';
import type { TournamentsRecordStatus } from './tournamentsRecordStatus.ts';

export interface TournamentsRecord {
  /** Unique record identifier */
  id?: string;
  title: string;
  slug?: string;
  description?: string;
  venue?: string;
  start_at?: string;
  end_at?: string;
  status: TournamentsRecordStatus;
  registration_enabled: boolean;
  registration_open_at?: string;
  registration_close_at?: string;
  max_teams?: number;
  min_team_size: number;
  max_team_size: number;
  bracket_count: number;
  bracket_format: TournamentsRecordBracketFormat;
  match_best_of: number;
  archived: boolean;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  created_by?: string;
  /** Relation field. When expanded, this becomes the related record object. Use the 'expand' query parameter to expand relations. */
  updated_by?: string;
  created?: string;
  updated?: string;
}